# perf(core): avoid extra then on opAsync()