@lucacasonato 2021-10-04
I am not really in favor, because stack traces become significantly worse.